requirements.txt.txt
Django==5.1.6
djangorestframework==3.14.0
requests==2.31.0